[REQ]22415 : To display image during preview
This change had to be done in order to display image in the preview window. Preview window is opened in a new app and it does not have any information about the core. Hence, the wrapper.url has to be stored in window's localstorage and retrieved whenever core is unavailable.